requirements.txt
cohere==5.14.0
firebase-admin==6.7.0
Flask==3.1.0
Flask-Session==0.8.0
google-auth==2.38.0
gunicorn==23.0.0
Jinja2==3.1.6
keras==3.9.0
numpy==2.1.3
pillow==11.1.0
python-dotenv==1.0.1
requests==2.32.3
tensorflow==2.19.0
Werkzeug==3.1.3
